refactor: move Position to own file
